+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `ls` command for cli
+- Response in cli if file does not exist
 
 
 ## [0.1.0] - 2023-08-07